+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Use PHP 7.1 (type hinting, strict types and etc)
 * Move classes to `src` dir
+* Remove native truncate stack feature (don't supported for PHP 7)
 
 ### 1.0.5 (02.02.2019)